backend/requirements.txt
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
-e git+https://github.com/yformer/EfficientSAM.git@d525f622e6f640acf5a0fc37c7ca1f243da5bde0#egg=efficient_sam
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
groundingdino-py==0.4.0
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.44.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segment-anything==1.0
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
supervision==0.6.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
umap==0.1.1
urllib3==2.5.0
uv==0.7.14
uvicorn==0.34.3
yapf==0.43.0